codebase/requirements.txt
# Core LLM Libraries
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# LangChain Framework
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20

# Vector Databases
chromadb>=0.4.22
pinecone-client>=3.0.0
faiss-cpu>=1.7.4

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
tiktoken>=0.5.2

# Web Framework
streamlit>=1.31.0
gradio>=4.19.0
flask>=3.0.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0
requests>=2.31.0
python-multipart>=0.0.9

# Data Processing
numpy>=1.26.0
pandas>=2.2.0

# Optional: For advanced features
redis>=5.0.0
celery>=5.3.0
langsmith>=0.0.87

# Development Tools
pytest>=8.0.0
black>=24.0.0
flake8>=7.0.0
